service/requirements.txt
xmltodict==0.11.0
requests==2.20.0
Flask==0.11
google-cloud-storage